requirements.txt
astroid==2.6.2
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
multitasking==0.0.9
numpy==1.21.0
pandas==1.3.0
pycodestyle==2.7.0
pylint==2.9.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
yfinance==0.1.60
